# api/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/articles":{"get":{"tags":["article-entity-controller","articles"],"summary":"Get articles","description":"Returns a list of articles.","operationId":"getCollectionResource-article-get_1_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}},{"name":"email","in":"header","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]}},{"name":"maxPrice","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"searchTerm","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Login success","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelArticle"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelArticle"}},"text/uri-list":{"schema":{"type":"string"}}}},"401":{"description":"Login failed","content":{"application/json":{}}}}},"post":{"tags":["article-entity-controller","articles"],"summary":"Create an article","description":"Creates an article.","operationId":"postCollectionResource-article-post_1","parameters":[{"name":"email","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArticleDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"200":{"description":"Article created","content":{"application/json":{}}}}}},"/articles/search/findByFilters":{"get":{"tags":["article-search-controller"],"operationId":"executeSearch-article-get","parameters":[{"name":"searchTerm","in":"query","schema":{"type":"string"}},{"name":"maxPrice","in":"query","schema":{"type":"number","format":"double"}},{"name":"articleStatus","in":"query","schema":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelArticle"}}}},"404":{"description":"Not Found"}}}},"/articles/search/findByStatusOrderByDescriptionAsc":{"get":{"tags":["article-search-controller"],"operationId":"executeSearch-article-get_1","parameters":[{"name":"status","in":"query","schema":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelArticle"}}}},"404":{"description":"Not Found"}}}},"/articles/{id}":{"get":{"tags":["article-entity-controller"],"description":"get-article","operationId":"getItemResource-article-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["article-entity-controller"],"description":"update-article","operationId":"putItemResource-article-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["article-entity-controller","articles"],"summary":"Delete an article","description":"Deletes an article.","operationId":"deleteItemResource-article-delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"email","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"},"200":{"description":"Article deleted","content":{"application/json":{}}}}},"patch":{"tags":["article-entity-controller"],"description":"patch-article","operationId":"patchItemResource-article-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"204":{"description":"No Content"}}}},"/articles/{id}/buyer":{"get":{"tags":["article-property-reference-controller"],"description":"get-customer-by-article-Id","operationId":"followPropertyReference-article-get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}},"text/uri-list":{"schema":{"type":"string"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["article-property-reference-controller"],"description":"update-customer-by-article-Id","operationId":"createPropertyReference-article-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["article-property-reference-controller"],"description":"delete-customer-by-article-Id","operationId":"deletePropertyReference-article-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["article-property-reference-controller"],"description":"patch-customer-by-article-Id","operationId":"createPropertyReference-article-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}}},"/articles/{id}/buyer/{propertyId}":{"get":{"tags":["article-property-reference-controller"],"description":"get-customer-by-article-Id","operationId":"followPropertyReference-article-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"404":{"description":"Not Found"}}},"delete":{"tags":["article-property-reference-controller"],"description":"delete-customer-by-article-Id","operationId":"deletePropertyReferenceId-article-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/articles/{id}/seller":{"get":{"tags":["article-property-reference-controller"],"description":"get-customer-by-article-Id","operationId":"followPropertyReference-article-get_2_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}},"text/uri-list":{"schema":{"type":"string"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["article-property-reference-controller"],"description":"update-customer-by-article-Id","operationId":"createPropertyReference-article-put_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["article-property-reference-controller"],"description":"delete-customer-by-article-Id","operationId":"deletePropertyReference-article-delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["article-property-reference-controller"],"description":"patch-customer-by-article-Id","operationId":"createPropertyReference-article-patch_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}}},"/articles/{id}/seller/{propertyId}":{"get":{"tags":["article-property-reference-controller"],"description":"get-customer-by-article-Id","operationId":"followPropertyReference-article-get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"404":{"description":"Not Found"}}},"delete":{"tags":["article-property-reference-controller"],"description":"delete-customer-by-article-Id","operationId":"deletePropertyReferenceId-article-delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/customers":{"get":{"tags":["customer-entity-controller"],"description":"get-customer","operationId":"getCollectionResource-customer-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelCustomer"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelCustomer"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["customer-entity-controller"],"description":"create-customer","operationId":"postCollectionResource-customer-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}}}}},"/customers/search/findByEmail":{"get":{"tags":["customer-search-controller"],"operationId":"executeSearch-customer-get","parameters":[{"name":"email","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"404":{"description":"Not Found"}}}},"/customers/{id}":{"get":{"tags":["customer-entity-controller"],"description":"get-customer","operationId":"getItemResource-customer-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["customer-entity-controller"],"description":"update-customer","operationId":"putItemResource-customer-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["customer-entity-controller"],"description":"delete-customer","operationId":"deleteItemResource-customer-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["customer-entity-controller"],"description":"patch-customer","operationId":"patchItemResource-customer-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"204":{"description":"No Content"}}}},"/customers/{id}/boughtArticles":{"get":{"tags":["customer-property-reference-controller"],"description":"get-article-by-customer-Id","operationId":"followPropertyReference-customer-get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}},"text/uri-list":{"schema":{"type":"string"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["customer-property-reference-controller"],"description":"update-article-by-customer-Id","operationId":"createPropertyReference-customer-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["customer-property-reference-controller"],"description":"delete-article-by-customer-Id","operationId":"deletePropertyReference-customer-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["customer-property-reference-controller"],"description":"patch-article-by-customer-Id","operationId":"createPropertyReference-customer-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"204":{"description":"No Content"}}}},"/customers/{id}/boughtArticles/{propertyId}":{"get":{"tags":["customer-property-reference-controller"],"description":"get-article-by-customer-Id","operationId":"followPropertyReference-customer-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"404":{"description":"Not Found"}}},"delete":{"tags":["customer-property-reference-controller"],"description":"delete-article-by-customer-Id","operationId":"deletePropertyReferenceId-customer-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/customers/{id}/soldArticles":{"get":{"tags":["customer-property-reference-controller"],"description":"get-article-by-customer-Id","operationId":"followPropertyReference-customer-get_2_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}},"text/uri-list":{"schema":{"type":"string"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["customer-property-reference-controller"],"description":"update-article-by-customer-Id","operationId":"createPropertyReference-customer-put_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["customer-property-reference-controller"],"description":"delete-article-by-customer-Id","operationId":"deletePropertyReference-customer-delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["customer-property-reference-controller"],"description":"patch-article-by-customer-Id","operationId":"createPropertyReference-customer-patch_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/CollectionModelObject"}},"text/uri-list":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"204":{"description":"No Content"}}}},"/customers/{id}/soldArticles/{propertyId}":{"get":{"tags":["customer-property-reference-controller"],"description":"get-article-by-customer-Id","operationId":"followPropertyReference-customer-get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/CollectionModelArticle"}}}},"404":{"description":"Not Found"}}},"delete":{"tags":["customer-property-reference-controller"],"description":"delete-article-by-customer-Id","operationId":"deletePropertyReferenceId-customer-delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"propertyId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}}},"/profile":{"get":{"tags":["profile-controller"],"operationId":"listAllFormsOfMetadata_1","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/RepresentationModelObject"}}}}}}},"/profile/articles":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/profile/customers":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/articles/{id}/auction":{"put":{"tags":["articles"],"summary":"Start or stop an auction","description":"Starts or stops an auction.","operationId":"startAuction","parameters":[{"name":"email","in":"header","required":true,"schema":{"type":"string"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"action","in":"query","required":true,"schema":{"type":"string","enum":["START","STOP"]}}],"responses":{"200":{"description":"Auction started","content":{"application/json":{}}}}}},"/bid":{"post":{"tags":["bid"],"summary":"Bid on an article","operationId":"bid","parameters":[{"name":"email","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BidDto"}}},"required":true},"responses":{"200":{"description":"Bid success","content":{"application/json":{}}},"400":{"description":"Bid failed","content":{"application/json":{}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Validates login credentials (Just kidding, we are just looking for the username).","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"401":{"description":"Login failed","content":{"application/json":{}}},"200":{"description":"Login success","content":{"application/json":{}}}}}}},"components":{"schemas":{"AbstractJsonSchemaPropertyObject":{"type":"object","properties":{"title":{"type":"string"},"readOnly":{"type":"boolean"}}},"Item":{"type":"object","properties":{"type":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}}}},"JsonSchema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}},"definitions":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Item"}},"type":{"type":"string"},"$schema":{"type":"string"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"RepresentationModelObject":{"type":"object","properties":{"_links":{"$ref":"#/components/schemas/Links"}}},"Address":{"type":"object","properties":{"street":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipCode":{"type":"string"},"country":{"type":"string"}}},"EntityModelCustomer":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"shippingAddress":{"$ref":"#/components/schemas/Address"},"paymentAddress":{"$ref":"#/components/schemas/Address"},"_links":{"$ref":"#/components/schemas/Links"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelCustomer":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"customers":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelCustomer"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"CollectionModelArticle":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"articles":{"type":"array","items":{"$ref":"#/components/schemas/ArticleResponse"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"CollectionModelObject":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"objects":{"type":"array","items":{"type":"object"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelArticle":{"type":"object","properties":{"description":{"type":"string"},"reservePrice":{"type":"number","format":"float"},"hammerPrice":{"type":"number","format":"double"},"auctionStartDate":{"type":"string","format":"date-time"},"auctionEndDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]},"_links":{"$ref":"#/components/schemas/Links"}}},"PagedModelEntityModelArticle":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"articles":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"CollectionModelEntityModelArticle":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"articles":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelArticle"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"ArticleRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"description":{"type":"string"},"reservePrice":{"type":"number","format":"float"},"hammerPrice":{"type":"number","format":"double"},"auctionStartDate":{"type":"string","format":"date-time"},"auctionEndDate":{"type":"string","format":"date-time"},"seller":{"type":"string"},"buyer":{"type":"string"},"status":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]},"bids":{"type":"array","items":{"type":"string"}},"categories":{"type":"array","items":{"type":"string"}}}},"CustomerRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"shippingAddress":{"$ref":"#/components/schemas/Address"},"paymentAddress":{"$ref":"#/components/schemas/Address"},"paymentOptions":{"type":"array","items":{"type":"string"}},"soldArticles":{"type":"array","items":{"type":"string"}},"boughtArticles":{"type":"array","items":{"type":"string"}},"bids":{"type":"array","items":{"type":"string"}}}},"ArticleResponse":{"type":"object","properties":{"description":{"type":"string"},"reservePrice":{"type":"number","format":"float"},"hammerPrice":{"type":"number","format":"double"},"auctionStartDate":{"type":"string","format":"date-time"},"auctionEndDate":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["LISTED","AUCTION_RUNNING","SOLD","NOT_SOLD"]}}},"BidDto":{"type":"object","properties":{"amount":{"type":"number","format":"double"},"bidder":{"type":"string"},"articleId":{"type":"integer","format":"int64"}}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"CreateArticleDto":{"type":"object","properties":{"description":{"type":"string"},"price":{"type":"number","format":"float"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}